backend/requirements.txt
# Core
fastapi
uvicorn[standard]
websockets
python-dotenv
pydantic
pydantic-settings

# Google Cloud
google-cloud-speech
google-cloud-texttospeech

# LLM
langchain
langgraph
langchain-openai
langgraph-checkpoint-sqlite
aiosqlite
# openai==1.6.1

# STT
assemblyai 
pyaudio

websockets 
sounddevice 
asyncio

# Audio TTS
elevenlabs
webrtcvad

# HTTP
httpx